requirements.txt
torch==1.10.1
torchvision
pycocotools
torch-lucent
captum
